stat: use gsl functions

+++ include/stats/stats.h
@@ -28,26 +28,26 @@ SOFTWARE
 /*
 Find the mean of an array
 */
-float stats_mean(float *arr, size_t len);
+double stats_mean(double *arr, size_t len);
 
 /*
 Find the variance
 */
-float stats_var(float *arr, size_t len);
+double stats_var(double *arr, size_t len);
 
 /*
 Find the covariance of two matrix
 */
-float stats_covar(float *arr1, float *arr2, size_t len);
+double stats_covar(double *arr1, double *arr2, size_t len);
 
 /*
 Returns the max value in the array
 */
-float stats_max(float *arr, size_t len);
+double stats_max(double *arr, size_t len);
 
 /*
 Returns the min value in the array
 */
-float stats_min(float *arr, size_t len);
+double stats_min(double *arr, size_t len);
 
 #endif
